Fix the tablet not displaying the Standing Stones required
Fix the Time Shift ritual being able to get activated with any incense instead of just clocks

+++ src/main/java/elucent/rootsclassic/ritual/rituals/RitualTimeShift.java
@@ -32,8 +32,8 @@ public class RitualTimeShift extends RitualBase {
     }
     inventory.clear();
     if (!world.isRemote && world.getServer() != null) {
-      for (ServerWorld serverworld : world.getServer().getWorlds()) {
-        serverworld.setDayTime(serverworld.getDayTime() + shiftAmount);
+      for(ServerWorld serverworld : world.getServer().getWorlds()) {
+        serverworld.setDayTime(serverworld.getDayTime() + (long)shiftAmount);
       }
     }
   }
